+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 * Add AutoCli project
 * Add CLI service, method, and parameter attributes
 * Add CLI class to map services
+* Add CLI mapping of extension methods
 * Add demo